requirements.txt
asgiref==3.9.1
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2
django-admin-ordering==0.20.0
django-autoslug==1.9.9
django-email-extras==0.3.4
django-filter==25.1
django-formtools==2.5.1
django-js-asset==3.1.2
django-json-widget==2.0.3
django-nine==0.2.7
django-nonefield==0.4
django-viewflow==2.0.2
dynamic-django-forms==0.1.11
easy-thumbnails==2.10
et_xmlfile==2.0.0
form-designer==0.27.0
future==0.15.0
idna==3.10
openpyxl==3.1.2
packaging==25.0
pillow==11.3.0
python-gnupg==0.5.5
pytz==2025.2
requests==2.32.4
six==1.17.0
sphinx-me==0.3
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
vishap==0.1.5
webencodings==0.5.1
xlsxdocument==1.3.10
